requirements.txt
apache-airflow==2.9.1
pandas==2.2.2
seaborn==0.13.2
scikit-learn==1.5.0